feature/confirm deletion (#28)
* feat(app): Ask user for confirmation before removing a schedule

* test(app): Update tests for 'InteractiveMode'

+++ src/core/modes/InteractiveMode.ts
@@ -135,6 +135,13 @@ class InteractiveMode extends Mode {
       this.scheduleService.schedules
     );
 
+    const isConfirmed = await this.prompt.getConfirmation(
+      `Are you sure you want to remove schedule '${scheduleToRemove}'?`
+    );
+    if (!isConfirmed) {
+      return this.logger.warn(`Action aborted.`);
+    }
+
     await this.deleteSchedule(scheduleToRemove);
   }
